Updated for akka 2.3.6 and stream/http version 0.7

+++ akka-http-example/src/main/scala/RandomDataPublisher.scala
@@ -31,8 +31,8 @@ class RandomDataPublisher extends ActorPublisher[ChunkStreamPart] with ActorLogg
     case _ =>
   }
 
-  def sendDataChunks(cnt: Int) {  
-    for(i <- 1 to cnt) {
+  def sendDataChunks(cnt: Long) {  
+    for(i <- 1L to cnt) {
       if(isActive && totalDemand > 0) {
         count += 1
         println(s"Sending Data! ($count)")
